+++ megalinter/reporters/TextReporter.py
@@ -63,8 +63,10 @@ class TextReporter(Reporter):
         ]
         text_report_lines += self.report_items
         text_report_lines += self.master.complete_text_reporter_report(self)
+        text_report_sub_folder = os.environ.get("TEXT_REPORTER_SUB_FOLDER", "text")
         text_file_name = (
             f"{self.report_folder}{os.path.sep}"
+            f"{text_report_sub_folder}{os.path.sep}"
             f"{self.master.status.upper()}-mega-linter-{self.master.name}.log"
         )
         if not os.path.isdir(os.path.dirname(text_file_name)):